requirements.txt
wheel
Faker
flask
flask-jwt
flask-jwt-extended
PyJWT
flask_cors
flask_restful
flask_sqlalchemy
gunicorn
jsonschema
marshmallow
marshmallow_sqlalchemy
psycopg2-binary
psycopg2
python-abc
python-dotenv
sqlalchemy
exceptiongroup
requests
Werkzeug
pytest
pytest-cov
pytest-mock
pytest-env